update: refactored PixelShuffle and UpScale
