+++ restorers/model/nafnet/__init__.py
@@ -1,2 +1,3 @@
-from .nafnet import NAFNet, PixelShuffle, UpScale
+from .nafnet import NAFNet
+from .blocks import PixelShuffle, UpScale
 from .nafblock import NAFBlock, SimpleGate, SimplifiedChannelAttention, ChannelAttention
